改字段类型

+++ dsp/src/main/java/com/tecdo/entity/AdGroup.java
@@ -53,7 +53,7 @@ public class AdGroup extends BaseEntity {
     /**
      * 竞价策略
      */
-    private Double bidStrategy;
+    private Integer bidStrategy;
 
     /**
      * 操作价格
